infra/requirements.txt
aws-cdk-lib==2.159.0
boto3==1.35.22
constructs>=10.0.0,<11.0.0
requests==2.32.*
